[IDLE-83] refactor: 패키지 구조 변경 - 4 layers 구조를 3 layers 구조로 변경 - 의전성 역전 법칙을 적용한 계층형 아키텍처로 변경
