+++ backend/book-service/src/main/java/kr/mybrary/bookservice/book/domain/PlatformBookSearchApiService.java
@@ -1,0 +1,8 @@
+package kr.mybrary.bookservice.book.domain;
+
+public interface PlatformBookSearchApiService {
+
+    Object searchWithKeyWord(String keyword);
+
+    Object searchWithISBN(String isbn);
+}
