+++ backend/book-service/src/main/java/kr/mybrary/bookservice/book/domain/dto/kakaoapi/Meta.java
@@ -1,4 +1,4 @@
-package kr.mybrary.bookservice.book.infrastructure.dto.kakaoapi;
+package kr.mybrary.bookservice.book.domain.dto.kakaoapi;
 
 import lombok.Data;
