--- backend/book-service/src/main/java/kr/mybrary/bookservice/book/infrastructure/booksearch/PlatformBookSearchApiRequester.java
@@ -1,8 +1,0 @@
-package kr.mybrary.bookservice.book.infrastructure.booksearch;
-
-public interface PlatformBookSearchApiRequester {
-
-    Object searchWithKeyWord(String keyword);
-
-    Object searchWithISBN(String isbn);
-}
